handle compressed content, some updates to unit-tests
