+++ mockstack/constants.py
@@ -12,6 +12,17 @@ PROXYRULES_FILE_TEMPLATE_PREFIX = "file:///"
 
 SENSITIVE_HEADERS = ["authorization", "cookie", "set-cookie"]
 
+# See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Encoding
+CONTENT_ENCODING_COMPRESSED = (
+    "gzip",
+    "compress",
+    "deflate",
+    "br",
+    "zstd",
+    "dcb",
+    "dcz",
+)
+
 
 class ProxyRulesRedirectVia(StrEnum):
     """The type of redirect to use for the proxy rules strategy.
