+++ mockstack/tests/strategies/test_proxyrules.py
@@ -13,141 +13,8 @@ from mockstack.constants import ProxyRulesRedirectVia
 from mockstack.strategies.proxyrules import (
     ProxyRulesStrategy,
     Rule,
-    URLRuleResult,
-    TemplateRuleResult,
+    maybe_update_headers,
 )
-
-
-def test_rule_from_dict():
-    """Test creating a Rule from a dictionary."""
-    data = {
-        "pattern": r"/api/v1/projects/(\d+)",
-        "replacement": r"/projects/\1",
-        "method": "GET",
-    }
-    rule = Rule.from_dict(data)
-    assert rule.pattern == data["pattern"]
-    assert rule.replacement == data["replacement"]
-    assert rule.method == data["method"]
-
-
-def test_rule_from_dict_without_method():
-    """Test creating a Rule from a dictionary without a method."""
-    data = {
-        "pattern": r"/api/v1/projects/(\d+)",
-        "replacement": r"/projects/\1",
-    }
-    rule = Rule.from_dict(data)
-    assert rule.pattern == data["pattern"]
-    assert rule.replacement == data["replacement"]
-    assert rule.method is None
-
-
-@pytest.mark.parametrize(
-    "pattern,path,method,expected",
-    [
-        (r"/api/v1/projects/\d+", "/api/v1/projects/123", "GET", True),
-        (r"/api/v1/projects/\d+", "/api/v1/projects/123", "POST", True),
-        (r"/api/v1/projects/\d+", "/api/v1/projects/abc", "GET", False),
-        (r"/api/v1/projects/\d+", "/api/v1/users/123", "GET", False),
-        (r"/api/v1/projects/\d+", "/api/v1/projects/123", "POST", True),
-    ],
-)
-def test_rule_matches(pattern, path, method, expected):
-    """Test the rule matching logic."""
-    rule = Rule(pattern=pattern, replacement="", method=None)
-    request = Request(
-        scope={
-            "type": "http",
-            "method": method,
-            "path": path,
-            "query_string": b"",
-            "headers": [],
-        }
-    )
-    assert rule.matches(request) == expected
-
-
-@pytest.mark.parametrize(
-    "pattern,path,method,expected",
-    [
-        (r"/api/v1/projects/\d+", "/api/v1/projects/123", "GET", True),
-        (r"/api/v1/projects/\d+", "/api/v1/projects/123", "POST", False),
-    ],
-)
-def test_rule_matches_with_method(pattern, path, method, expected):
-    """Test the rule matching logic with method restriction."""
-    rule = Rule(pattern=pattern, replacement="", method="GET")
-    request = Request(
-        scope={
-            "type": "http",
-            "method": method,
-            "path": path,
-            "query_string": b"",
-            "headers": [],
-        }
-    )
-    assert rule.matches(request) == expected
-
-
-@pytest.mark.parametrize(
-    "pattern,replacement,path,expected_url",
-    [
-        (
-            r"/api/v1/projects/(\d+)",
-            r"/projects/\1",
-            "/api/v1/projects/123",
-            "/projects/123",
-        ),
-        (
-            r"/api/v1/users/([^/]+)",
-            r"/users/\1",
-            "/api/v1/users/john",
-            "/users/john",
-        ),
-    ],
-)
-def test_rule_apply(pattern, replacement, path, expected_url):
-    """Test the rule application logic."""
-    rule = Rule(pattern=pattern, replacement=replacement)
-    request = Request(
-        scope={
-            "type": "http",
-            "method": "GET",
-            "path": path,
-            "query_string": b"",
-            "headers": [],
-        }
-    )
-    result = rule.apply(request)
-    assert isinstance(result, URLRuleResult)
-    assert result.get_result_type() == "url"
-    assert result.url == expected_url
-
-
-def test_rule_apply_template():
-    """Test the rule application logic for template files."""
-    rule = Rule(
-        pattern=r"/api/v1/projects/(\d+)",
-        replacement=r"file:///path/to/template.json",
-    )
-    request = Request(
-        scope={
-            "type": "http",
-            "method": "GET",
-            "path": "/api/v1/projects/1234",
-            "query_string": b"",
-            "headers": [],
-        }
-    )
-    result = rule.apply(request)
-    assert isinstance(result, TemplateRuleResult)
-    assert result.get_result_type() == "template"
-    assert result.template_path == "/path/to/template.json"
-    assert result.template_context is not None
-    # The context should contain the extracted project ID from the path
-    assert "projects" in result.template_context
-    assert result.template_context["projects"] == "1234"
 
 
 def test_proxy_rules_strategy_load_rules(settings):
@@ -491,3 +358,16 @@ async def test_proxy_rules_strategy_reverse_proxy(settings_reverse_proxy, span):
         assert isinstance(headers, Headers)
 
         mock_client.send.assert_called_once()
+
+    def test_maybe_update_headers_updates_content_encoding():
+        """Test maybe_update_headers updates content-encoding."""
+        headers = Headers(
+            {"content-encoding": "gzip", "content-type": "application/json"}
+        )
+        request_headers = Headers({"accept-encoding": "gzip"})
+
+        updated_headers = maybe_update_headers(headers, request_headers)
+
+        assert updated_headers["content-encoding"] == "identity"
+        assert updated_headers["content-type"] == "application/json"
+        assert updated_headers["accept-encoding"] == "gzip"
